requirements.txt
flask==3.1.3
PyJWT==2.12.1
psycopg2-binary==2.9.10
